src/requirements.txt
fastapi
pydantic
jinja2
python-multipart
uvicorn
pyjwt
bcrypt
ruff
